requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-common-helpers==0.9.2
django-cron==0.5.1
django-filter==2.1.0
django-phone-field==1.7.1
djangorestframework==3.9.2
docopt==0.6.2
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
Markdown==3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.17
pipreqs==0.4.9
protobuf==3.9.1
Pygments==2.3.1
pylint==2.3.1
pytz==2019.1
requests==2.21.0
rope==0.14.0
six==1.12.0
sqlparse==0.3.0
-e git://github.com/tonyg/python-rfc3339.git@ed7ebfc3c6ec80cc0bb4708e7d44eaa02148472d#egg=tonyg_rfc3339
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.24.2
wrapt==1.11.2
yarg==0.1.9
-e git+https://gitlab.int/carrierx/logging_utils.git#egg=logging_utils